게시물 API 수정

+++ src/main/java/com/comesfullcircle/board/service/PostService.java
@@ -1,6 +1,8 @@
 package com.comesfullcircle.board.service;
 
 import com.comesfullcircle.board.exception.post.PostNotFoundException;
+import com.comesfullcircle.board.exception.user.UserNotAllowedException;
+import com.comesfullcircle.board.model.entity.UserEntity;
 import com.comesfullcircle.board.model.post.Post;
 import com.comesfullcircle.board.model.post.PostPatchRequestBody;
 import com.comesfullcircle.board.model.post.PostPostRequestBody;
@@ -31,24 +33,34 @@ public class PostService {
         return Post.from(postEntity);
     }
 
-    public Post createPost(PostPostRequestBody postPostRequestBody) {
-        var postEntity = new PostEntity();
-        postEntity.setBody(postPostRequestBody.body());
-        var savedPostEntity = postEntityRepository.save(postEntity);
-        return Post.from(savedPostEntity);
+    public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
+        var postEntity = postEntityRepository.save(
+                PostEntity.of(postPostRequestBody.body(), currentUser)
+        );
+        return Post.from(postEntity);
     }
 
-    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody) {
+    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody, UserEntity currentUser) {
         var postEntity = postEntityRepository.findById(postId)
                 .orElseThrow(() -> new PostNotFoundException(postId));
+
+        if (!postEntity.getUser().equals(currentUser)) {
+            throw new UserNotAllowedException();
+        }
+
         postEntity.setBody(postPatchRequestBody.body());
         var updatedPostEntity = postEntityRepository.save(postEntity);
         return Post.from(updatedPostEntity);
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, UserEntity currentUser) {
         var postEntity = postEntityRepository.findById(postId)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Post not found"));
+                .orElseThrow(() -> new PostNotFoundException(postId));
+
+        if (!postEntity.getUser().equals(currentUser)) {
+            throw new UserNotAllowedException();
+        }
+
         postEntityRepository.delete(postEntity);
     }
 }
